+++ gateway-service/internal/model/prediction.go
@@ -6,6 +6,6 @@ type Prediction struct {
 	ID         string
 	DocumentID string
 	Label      string
-	Confidence float32
+	Confidence float64
 	CreatedAt  time.Time
 }
